requirements.txt
click==6.7
dominate==2.3.1
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
MarkupSafe==1.0
numpy==1.13.3
openpyxl==2.4.9
pandas==0.21.1
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
visitor==0.1.3
Werkzeug==0.13
python-telegram-bot
